Make Validate->Execute DDL identity and Oracle object identity canonical
Execute recomputed the Map->DDL fingerprint from live catalog spellings while
Validate had stamped it from Map stamps, so a bound existing table refused
every run with "DDL identity mismatch - No rows committed" on a green
Validate. Identity is now the *materialized* destination DDL (transform lives
in the conversion/risk contract, not physical identity), and the hash the
operator approved at Validate is carried UI->API into Execute, so the gate
still refuses genuine drift instead of being disarmed.

Same root cause silently broke Gate-8: identical rows hashed differently when
the two sides spelled a type `bigint` vs `BIGINT` or carried a collation.

Oracle: a quoted lower-case table and its folded upper-case twin are two
objects, and writer / introspection / collision probe / Gate-8 read-back each
resolved the name their own way - an append into an existing "users" created
a second USERS and split the job's rows across both while both runs passed.
One resolver (services/sql_object_identity.py) now answers what the object is
actually called, used by all four. The duplicate probe additionally compared
keys as CLOB (ORA-22849) and degraded to a skip, so duplicates were only
caught by the database at write time; it now casts to a bounded VARCHAR2
carrier and keeps Text for keys wider than 4000 rather than truncating.

Extraction (module budget): per-engine sample reads -> services/target_sample.py,
Oracle read-back -> services/reconciliation_oracle.py.

+++ apps/api/tests/test_conversion_contract_module12.py
@@ -139,11 +139,13 @@ def test_assert_ddl_identity_fail_closed_on_drift():
     fp = approved_mapping_ddl_fingerprint(maps, dest_db="postgresql")
     assert assert_ddl_identity(fp, maps, dest_db="postgresql") == fp
 
+    # INTEGER and BIGINT materialize to the same Postgres column under the
+    # never-narrower rule; SMALLINT is a different physical column.
     drifted = [
         {
             "source": "id",
             "target": "id",
-            "target_type": "BIGINT",
+            "target_type": "SMALLINT",
             "transform": "none",
         }
     ]
